refactor: 사다리를 생성할 때 List가 아닌 Ladder를 직접 생성하도록 변경
